# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Zabka Go API
  description: Backend projektu "Żabka Go" na zaliczenie przedmiotu "Aplikacje Mobilne"
  version: 0.0.1
servers:
  - url: http://127.0.0.1:5000/v1/
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: session_token
...
paths:
...
  /shop/{shop_id}/comment:
    post:
      summary: Makes comment in shop discussion.
      tags:
        - Comment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: shop_id
          required: true
          schema:
            type: integer
        - in: query
          name: content
          required: true
          schema:
            type: string
        - in: query
          name: parent_id
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Comment addition was successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
              example:
                status: "success"
        '400':
          description: Occurs when parent comment belongs to different shop discussion than shop with specified id.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
              example:
                message: "shop_id_and_parent_id_mismatch"
                status: "fail"
        '401':
          description: Token session was invalid or not provided.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
              example:
                message: "no_session_token"
                status: "fail"
        '403':
          description: User is not allowed to comment because he didn't visit the shop.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
              example:
                message: "shop_not_visited"
                status: "fail"
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
              example:
                message: "shop_not_found"
                status: "fail"
  /shop/{shop_id}/comments:
    get:
      summary: Returns all comments in shop discussion.
      tags:
        - Comment
      parameters:
        - in: path
          name: shop_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Users data was returned with no issues.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    parent_id:
                      type: integer
                    place_id:
                      type: integer
                    content:
                      type: string
                    creation_time:
                      type: string
                      format: date
              example:
                - content: "Testowy komentarz"
                  creation_time: "Tue, 07 May 2024 13:16:00 GMT"
                  id: 1
                  parent_id: 1
                  place_id: 4
                  user_id: 2
                - content: "Testowy podkomentarz"
                  creation_time: "Tue, 07 May 2024 13:21:53 GMT"
                  id: 2
                  parent_id: 1
                  place_id: 4
                  user_id: 2
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
              example:
                message: "shop_not_found"
                status: "fail"
  /shop/{shop_id}/comment/{comment_id}:
    get:
      summary: Return data of single comment in shop discussion.
      tags:
        - Comment
      parameters:
        - in: path
          name: shop_id
          required: true
          schema:
            type: integer
        - in: path
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Users data was returned with no issues.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  parent_id:
                    type: integer
                  place_id:
                    type: integer
                  content:
                    type: string
                  creation_time:
                    type: string
                    format: date
              example:
                content: "Testowy komentarz"
                creation_time: "Tue, 07 May 2024 13:16:00 GMT"
                id: 1
                parent_id: 1
                place_id: 4
                user_id: 2
        '404':
          description: Shop or comment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
              example:
                message: "comment_not_found"
                status: "fail"
...